SIMD: Make use of AVX intrinsics. Initial polyphony support.

+++ processor/processor_matrix.cpp
@@ -244,7 +244,7 @@ void ProcessorMatrix::set_unlock()
 }
 
 
-double ProcessorMatrix::process()
+float8 ProcessorMatrix::process()
 {
     for (auto& i : processors)
         i->process();
